+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add gitee mirror
 - Add RDA8910 chip flasher (Air724)
 - Fix ESP32C3,ESP32S3 flasher error
+- Add QS100 chip flasher